+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -44,6 +44,7 @@ import at.hannibal2.skyhanni.features.misc.*
 import at.hannibal2.skyhanni.features.misc.discordrpc.DiscordRPCManager
 import at.hannibal2.skyhanni.features.misc.items.EstimatedItemValue
 import at.hannibal2.skyhanni.features.misc.items.EstimatedWardrobePrice
+import at.hannibal2.skyhanni.features.misc.tabcomplete.WarpTabComplete
 import at.hannibal2.skyhanni.features.misc.teleportpad.TeleportPadCompactName
 import at.hannibal2.skyhanni.features.misc.teleportpad.TeleportPadInventoryNumber
 import at.hannibal2.skyhanni.features.misc.tiarelay.TiaRelayHelper
@@ -142,6 +143,8 @@ class SkyHanniMod {
         loadModule(GardenAPI)
         loadModule(CollectionAPI())
         loadModule(FarmingContestAPI)
+        loadModule(FriendAPI())
+        loadModule(PartyAPI())
 
         // features
         loadModule(BazaarOrderHelper())
@@ -286,6 +289,7 @@ class SkyHanniMod {
         loadModule(CityProjectFeatures())
         loadModule(GardenPlotIcon)
         loadModule(ShowFishingItemName())
+        loadModule(WarpTabComplete)
 
         init()